Allow inclusion by relative path
Before this patch, the inclusion of other files was possible either by
specifying the path from the main file's directory, or by having defined the
variable SRCTREE with the path to the included files.

This patch introduces a third case, where the included file path is relative
to the calling config file's path. This simplifies the recursive inclusion of
configuration files.